Finshed init for sender.cpp and fixed msg.h warning

+++ Assignment_2/samplefiles/msg.h
@@ -23,6 +23,6 @@ struct message
  	 */
 	void print(FILE* fp)
 	{
-		fprintf(fp, "%d %d", mtype, size);
+		fprintf(fp, "%d %d", (int)mtype, size);
 	}
 };
